Remove CreatedBy tag
Remove global `CreatedBy` tag from every resource. This was intended
to store the user who ran Terraform to create the resource, but it was
actually updated every time we ran Terraform, causing a lot of churn
and overwriting the original value.

We couldn't find a way to add the tag on resource creation and freeze
it, so it seems better to remove it.

Terraform will eventually be run by a Terraform user in CI, so the
`CreatedBy` tag wouldn't be so useful anyway.