Add GraphQL API endpoint for backend file checks to use
Add new API endpoint to be used by the file check jobs. This is a separate endpoint
to the existing endpoint used by the frontend because the authentication config is
different: the frontend uses Cognito ID tokens, and the backend will use IAM user
keys to sign requests. You cannot add both types of auth to the same endpoint.