Remove SSM write permissions from DB migration task
Restrict the DB migration task's permissions so that it can only
read data from the parameter store, not write it. The task only
needs to read the DB connection parameters - it should never update
the parameter store.

This also means that we have to explicitly add permission to write
logs, since the SSM write permission included log writing.